refactor + remove executor impl
