+++ service/history/task/cross_cluster_source_task_executor.go
@@ -22,443 +22,30 @@
 package task
 
 import (
-	"context"
-	"errors"
-
-	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/log"
-	"github.com/uber/cadence/common/metrics"
-	"github.com/uber/cadence/common/persistence"
-	ctask "github.com/uber/cadence/common/task"
-	"github.com/uber/cadence/common/types"
 	"github.com/uber/cadence/service/history/config"
 	"github.com/uber/cadence/service/history/execution"
 	"github.com/uber/cadence/service/history/shard"
 )
 
-var (
-	errContinueExecution = errors.New("cross cluster task should continue execution")
-)
-
 type (
-	crossClusterSourceTaskExecutor struct {
-		shard          shard.Context
-		executionCache *execution.Cache
-		logger         log.Logger
-		metricsClient  metrics.Client
-		config         *config.Config
-	}
+	crossClusterSourceTaskExecutor struct{}
 )
 
+// NewCrossClusterSourceTaskExecutor creates a new executor for processing
+// cross cluster tasks at source cluster
 func NewCrossClusterSourceTaskExecutor(
 	shard shard.Context,
 	executionCache *execution.Cache,
 	logger log.Logger,
 	config *config.Config,
 ) Executor {
-	return &crossClusterSourceTaskExecutor{
-		shard:          shard,
-		executionCache: executionCache,
-		logger:         logger,
-		metricsClient:  shard.GetMetricsClient(),
-		config:         config,
-	}
+	return &crossClusterSourceTaskExecutor{}
 }
 
 func (t *crossClusterSourceTaskExecutor) Execute(
 	task Task,
 	shouldProcessTask bool,
 ) error {
-	sourceTask, ok := task.(*crossClusterSourceTask)
-	if !ok {
-		return errUnexpectedTask
-	}
-
-	if !shouldProcessTask {
-		// this should not happen for cross cluster task
-		// shouldProcessTask parameter will be deprecated after
-		// 3+DC task lifecycle is done.
-		return errors.New("encounter shouldProcessTask equals to false when processing cross cluster task at source cluster")
-	}
-
-	if err := t.verifyDomainActive(sourceTask); err != nil {
-		return err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), taskDefaultTimeout)
-	defer cancel()
-
-	var err error
-	switch task.GetTaskType() {
-	case persistence.CrossClusterTaskTypeStartChildExecution:
-		err = t.executeStartChildExecutionTask(ctx, sourceTask)
-	case persistence.CrossClusterTaskTypeCancelExecution:
-		err = t.executeCancelExecutionTask(ctx, sourceTask)
-	case persistence.CrossClusterTaskTypeSignalExecution:
-		err = t.executeSignalExecutionTask(ctx, sourceTask)
-	// TODO: implement the execution logic for recordChildCompleted and applyParentClosePolicy task
-	default:
-		err = errUnknownCrossClusterTask
-	}
-
-	if err == nil {
-		t.setTaskState(sourceTask, ctask.TaskStateAcked, processingStateResponseReported)
-	} else if err == errContinueExecution {
-		err = nil
-	}
-
-	return err
-}
-
-func (t *crossClusterSourceTaskExecutor) executeStartChildExecutionTask(
-	ctx context.Context,
-	task *crossClusterSourceTask,
-) (retError error) {
-
-	taskInfo := task.GetInfo().(*persistence.CrossClusterTaskInfo)
-	wfContext, mutableState, release, err := loadWorkflowForCrossClusterTask(ctx, t.executionCache, taskInfo, t.metricsClient, t.logger)
-	if err != nil || mutableState == nil {
-		return err
-	}
-	defer func() { release(retError) }()
-
-	initiatedEventID := taskInfo.ScheduleID
-	childInfo, ok := mutableState.GetChildExecutionInfo(initiatedEventID)
-	if !ok {
-		return nil
-	}
-
-	ok, err = verifyTaskVersion(t.shard, t.logger, taskInfo.DomainID, childInfo.Version, taskInfo.Version, task)
-	if err != nil || !ok {
-		return err
-	}
-
-	initiatedEvent, err := mutableState.GetChildExecutionInitiatedEvent(ctx, initiatedEventID)
-	if err != nil {
-		return err
-	}
-
-	if task.ProcessingState() == processingStateInvalidated {
-		return t.generateNewTask(ctx, wfContext, mutableState, task)
-	}
-
-	// handle common errors
-	failedCause := task.response.FailedCause
-	if failedCause != nil {
-		switch *failedCause {
-		case types.CrossClusterTaskFailedCauseDomainNotActive:
-			return t.generateNewTask(ctx, wfContext, mutableState, task)
-		case types.CrossClusterTaskFailedCauseWorkflowNotExists:
-			return &types.EntityNotExistsError{}
-		case types.CrossClusterTaskFailedCauseWorkflowAlreadyCompleted:
-			return &types.WorkflowExecutionAlreadyCompletedError{}
-		case types.CrossClusterTaskFailedCauseUncategorized:
-			// ideally this case should not happen
-			// crossClusterSourceTask.Update() will reject response with this failed cause
-			t.setTaskState(task, ctask.TaskStatePending, processingState(task.response.TaskState))
-			return errContinueExecution
-		}
-	}
-
-	switch processingState(task.response.TaskState) {
-	case processingStateInitialized:
-		if childInfo.StartedID != common.EmptyEventID {
-			// TODO: explain when this may happen
-			t.setTaskState(task, ctask.TaskStatePending, processingStateResponseRecorded)
-			return errContinueExecution
-		}
-
-		attributes := initiatedEvent.StartChildWorkflowExecutionInitiatedEventAttributes
-		now := t.shard.GetTimeSource().Now()
-		if failedCause != nil &&
-			(*failedCause == types.CrossClusterTaskFailedCauseDomainNotExists ||
-				*failedCause == types.CrossClusterTaskFailedCauseWorkflowAlreadyRunning) {
-			return recordStartChildExecutionFailed(ctx, taskInfo, wfContext, attributes, now)
-		}
-
-		childRunID := task.response.StartChildExecutionAttributes.GetRunID()
-		err := recordChildExecutionStarted(ctx, taskInfo, wfContext, attributes, childRunID, now)
-		if err != nil {
-			return err
-		}
-
-		// advance the task state so that it will be available for polling again
-		t.setTaskState(task, ctask.TaskStatePending, processingStateResponseRecorded)
-		return errContinueExecution
-	case processingStateResponseRecorded:
-		// there's nothing we need to do for the remaining two failed causes
-		// ack the task
-		return nil
-	default:
-		return errUnknownTaskProcessingState
-	}
-}
-
-func (t *crossClusterSourceTaskExecutor) executeCancelExecutionTask(
-	ctx context.Context,
-	task *crossClusterSourceTask,
-) (retError error) {
-
-	taskInfo := task.GetInfo().(*persistence.CrossClusterTaskInfo)
-	wfContext, mutableState, release, err := loadWorkflowForCrossClusterTask(ctx, t.executionCache, taskInfo, t.metricsClient, t.logger)
-	if err != nil || mutableState == nil {
-		return err
-	}
-	defer func() { release(retError) }()
-
-	initiatedEventID := taskInfo.ScheduleID
-	requestCancelInfo, ok := mutableState.GetRequestCancelInfo(initiatedEventID)
-	if !ok {
-		return nil
-	}
-	ok, err = verifyTaskVersion(t.shard, t.logger, taskInfo.DomainID, requestCancelInfo.Version, taskInfo.Version, task)
-	if err != nil || !ok {
-		return err
-	}
-
-	if task.ProcessingState() == processingStateInvalidated {
-		return t.generateNewTask(ctx, wfContext, mutableState, task)
-	}
-
-	// handle common errors
-	failedCause := task.response.FailedCause
-	if failedCause != nil {
-		switch *failedCause {
-		case types.CrossClusterTaskFailedCauseDomainNotActive:
-			return t.generateNewTask(ctx, wfContext, mutableState, task)
-		case types.CrossClusterTaskFailedCauseWorkflowAlreadyRunning,
-			types.CrossClusterTaskFailedCauseUncategorized:
-			// ideally this case should not happen
-			// crossClusterSourceTask.Update() will reject response with those failed causes
-			t.setTaskState(task, ctask.TaskStatePending, processingState(task.response.TaskState))
-			return errContinueExecution
-		}
-	}
-
-	switch processingState(task.response.TaskState) {
-	case processingStateInitialized:
-		now := t.shard.GetTimeSource().Now()
-		targetDomainName, err := t.shard.GetDomainCache().GetDomainName(taskInfo.TargetDomainID)
-		if err != nil {
-			return err
-		}
-
-		if failedCause != nil {
-			// remaining errors are non-retryable
-			return requestCancelExternalExecutionFailed(
-				ctx,
-				taskInfo,
-				wfContext,
-				targetDomainName,
-				taskInfo.TargetWorkflowID,
-				taskInfo.TargetRunID,
-				now,
-			)
-		}
-		return requestCancelExternalExecutionCompleted(
-			ctx,
-			taskInfo,
-			wfContext,
-			targetDomainName,
-			taskInfo.TargetWorkflowID,
-			taskInfo.TargetRunID,
-			now,
-		)
-
-	default:
-		return errUnknownTaskProcessingState
-	}
-}
-
-func (t *crossClusterSourceTaskExecutor) executeSignalExecutionTask(
-	ctx context.Context,
-	task *crossClusterSourceTask,
-) (retError error) {
-
-	taskInfo := task.GetInfo().(*persistence.CrossClusterTaskInfo)
-	wfContext, mutableState, release, err := loadWorkflowForCrossClusterTask(ctx, t.executionCache, taskInfo, t.metricsClient, t.logger)
-	if err != nil || mutableState == nil {
-		return err
-	}
-	defer func() { release(retError) }()
-
-	initiatedEventID := taskInfo.ScheduleID
-	signalInfo, ok := mutableState.GetSignalInfo(initiatedEventID)
-	if !ok {
-		// TODO: here we should also RemoveSignalMutableState from target workflow
-		// Otherwise, target SignalRequestID still can leak if shard restart after signalExternalExecutionCompleted
-		// To do that, probably need to add the SignalRequestID in task info
-		return nil
-	}
-	ok, err = verifyTaskVersion(t.shard, t.logger, taskInfo.DomainID, signalInfo.Version, taskInfo.Version, task)
-	if err != nil || !ok {
-		return err
-	}
-
-	if task.ProcessingState() == processingStateInvalidated {
-		return t.generateNewTask(ctx, wfContext, mutableState, task)
-	}
-
-	// handle common errors
-	failedCause := task.response.FailedCause
-	if failedCause != nil {
-		switch *failedCause {
-		case types.CrossClusterTaskFailedCauseDomainNotActive:
-			return t.generateNewTask(ctx, wfContext, mutableState, task)
-		case types.CrossClusterTaskFailedCauseWorkflowAlreadyRunning,
-			types.CrossClusterTaskFailedCauseUncategorized:
-			// ideally this case should not happen
-			// crossClusterSourceTask.Update() will reject response with those failed causes
-			t.setTaskState(task, ctask.TaskStatePending, processingState(task.response.TaskState))
-			return errContinueExecution
-		}
-	}
-
-	switch processingState(task.response.TaskState) {
-	case processingStateInitialized:
-		now := t.shard.GetTimeSource().Now()
-		targetDomainName, err := t.shard.GetDomainCache().GetDomainName(taskInfo.TargetDomainID)
-		if err != nil {
-			return err
-		}
-
-		if failedCause != nil {
-			// remaining errors are non-retryable
-			return signalExternalExecutionFailed(
-				ctx,
-				taskInfo,
-				wfContext,
-				targetDomainName,
-				taskInfo.TargetWorkflowID,
-				taskInfo.TargetRunID,
-				signalInfo.Control,
-				now,
-			)
-		}
-
-		if err := signalExternalExecutionCompleted(
-			ctx,
-			taskInfo,
-			wfContext,
-			targetDomainName,
-			taskInfo.TargetWorkflowID,
-			taskInfo.TargetRunID,
-			signalInfo.Control,
-			now,
-		); err != nil {
-			return err
-		}
-
-		// advance the task state so that it will be available for polling again
-		t.setTaskState(task, ctask.TaskStatePending, processingStateResponseRecorded)
-		return errContinueExecution
-	case processingStateResponseRecorded:
-		// there's nothing we need to do for the response or remaining failed cause
-		// ack the task
-		return nil
-	default:
-		return errUnknownTaskProcessingState
-	}
-}
-
-func loadWorkflowForCrossClusterTask(
-	ctx context.Context,
-	executionCache *execution.Cache,
-	taskInfo *persistence.CrossClusterTaskInfo,
-	metricsClient metrics.Client,
-	logger log.Logger,
-) (execution.Context, execution.MutableState, execution.ReleaseFunc, error) {
-	wfContext, release, err := executionCache.GetOrCreateWorkflowExecutionWithTimeout(
-		taskInfo.GetDomainID(),
-		getWorkflowExecution(taskInfo),
-		taskGetExecutionContextTimeout,
-	)
-	if err != nil {
-		if err == context.DeadlineExceeded {
-			return nil, nil, nil, errWorkflowBusy
-		}
-		return nil, nil, nil, err
-	}
-
-	mutableState, err := loadMutableStateForCrossClusterTask(ctx, wfContext, taskInfo, metricsClient, logger)
-	if err != nil {
-		release(err)
-		return nil, nil, nil, err
-	}
-	if mutableState == nil || !mutableState.IsWorkflowExecutionRunning() {
-		release(nil)
-		return nil, nil, nil, nil
-	}
-
-	return wfContext, mutableState, release, nil
-}
-
-func (t *crossClusterSourceTaskExecutor) generateNewTask(
-	ctx context.Context,
-	wfContext execution.Context,
-	mutableState execution.MutableState,
-	task *crossClusterSourceTask,
-) error {
-	clusterMetadata := t.shard.GetClusterMetadata()
-	taskGenerator := execution.NewMutableStateTaskGenerator(
-		clusterMetadata,
-		t.shard.GetDomainCache(),
-		t.logger,
-		mutableState,
-	)
-	taskInfo := task.GetInfo().(*persistence.CrossClusterTaskInfo)
-	if err := taskGenerator.GenerateFromCrossClusterTask(taskInfo); err != nil {
-		return err
-	}
-
-	now := t.shard.GetTimeSource().Now()
-	isActive := clusterMetadata.ClusterNameForFailoverVersion(mutableState.GetCurrentVersion()) ==
-		clusterMetadata.GetCurrentClusterName()
-
-	var err error
-	if isActive {
-		err = wfContext.UpdateWorkflowExecutionAsActive(ctx, now)
-	} else {
-		err = wfContext.UpdateWorkflowExecutionAsPassive(ctx, now)
-	}
-	if err != nil {
-		return err
-	}
-
-	t.setTaskState(task, ctask.TaskStateAcked, processingStateInvalidated)
-	return nil
-}
-
-func (t *crossClusterSourceTaskExecutor) verifyDomainActive(
-	task *crossClusterSourceTask,
-) error {
-	entry, err := t.shard.GetDomainCache().GetDomainByID(task.GetDomainID())
-	if err != nil {
-		return err
-	}
-
-	if entry.IsDomainPendingActive() {
-		// return error so that the task can be retried
-		return ErrTaskPendingActive
-	}
-
-	if !entry.IsDomainActive() {
-		// set processing state to invalidated so that a new task can be created
-		t.setTaskState(task, ctask.TaskStatePending, processingStateInvalidated)
-		return nil
-	}
-
-	return nil
-}
-
-func (t *crossClusterSourceTaskExecutor) setTaskState(
-	task *crossClusterSourceTask,
-	state ctask.State,
-	processingState processingState,
-) {
-	task.Lock()
-	task.state = state
-	task.processingState = processingState
-	task.response = nil
-	task.Unlock()
+	panic("crossClusterSourceTaskExecutor not implemented")
 }
